Changa type fields `cost` and `price` to `FloatField`

+++ api/models.py
@@ -80,8 +80,8 @@ class BaseElement(models.Model):
     title = models.CharField(verbose_name="Название", max_length=160)
     measure = models.CharField(verbose_name="Единицы измерения", max_length=30)
     second_measure = models.CharField(verbose_name="Доп. ед. измерения", max_length=30)
-    cost = models.PositiveIntegerField(verbose_name="Себестоимость", default=0)
-    price = models.PositiveIntegerField(verbose_name="Цена", default=0)
+    cost = models.FloatField(verbose_name="Себестоимость", default=0)
+    price = models.FloatField(verbose_name="Цена", default=0)
     type = models.CharField(verbose_name="Тип", choices=Type.choices, max_length=30)
     dimension = models.CharField(verbose_name="Размер", max_length=60, blank=True)
     conversion_rate = models.FloatField(verbose_name="Норма конвертации")
@@ -143,8 +143,8 @@ class BaseConstruction(models.Model):
     title = models.CharField(verbose_name="Название", max_length=160)
     description = models.TextField(verbose_name="Описание", blank=True)
     measure = models.CharField(verbose_name="Единицы измерения", max_length=30)
-    cost = models.PositiveIntegerField(verbose_name="Себестоимость", default=0)
-    price = models.PositiveIntegerField(verbose_name="Цена", default=0)
+    cost = models.FloatField(verbose_name="Себестоимость", default=0)
+    price = models.FloatField(verbose_name="Цена", default=0)
 
     class Meta:
         abstract = True
